+++ NEWS.md
@@ -1,3 +1,8 @@
+# governor 0.1.0.9000  2024-07-09
+
+* Use `Sleep()` on Windows
+
+
 # governor 0.1.0  2024-06-29
 
 * Initial release